rename product actions
