+++ frontend/components/category/Category.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { listCategoryProducts } from "../../actions/categoryActions";
-import { fetchProducts } from "../../actions/product_actions";
+import { fetchProducts } from "../../actions/productActions";
 import Container from "../../modules/container/Container"
 import ProductItem from "../../modules/list/ProductItem";
 import "./Category.scss";
